Removes duplicate code in AnormDB
AnormDB had two similar implementations, SpanStore and Storage/Index.
The latter is a legacy, but had more recent maintenance than the former.

This change migrates the most recent code into AnormDB's SpanStore and
removes the code for Storage/Index.

See #483
