+++ zipkin-anormdb/src/main/scala/com/twitter/zipkin/anormdb/AnormDBSpanStoreFactory.scala
@@ -17,14 +17,14 @@ package com.twitter.zipkin.anormdb
 
 import com.twitter.app.App
 import com.twitter.zipkin.storage.SpanStore
-import com.twitter.zipkin.storage.anormdb.{AnormSpanStore, SpanStoreDB}
+import com.twitter.zipkin.storage.anormdb.{DB, DBConfig, AnormSpanStore}
 
 trait AnormDBSpanStoreFactory { self: App =>
-  val anormDB = flag("zipkin.storage.anormdb.db", "sqlite::memory:", "JDBC location URL for the AnormDB")
+  val anormDB = flag("zipkin.storage.anormdb.db", "sqlite-memory", "name The database type")
   val anormInstall = flag("zipkin.storage.anormdb.install", true, "Create the tables")
 
   def newAnormSpanStore(): SpanStore = {
-    val db = SpanStoreDB(anormDB())
+    val db = DB(new DBConfig(name = anormDB(), install = anormInstall()))
     val conn = if (anormInstall()) Some(db.install()) else None
     new AnormSpanStore(db, conn)
   }
