--- zipkin-anormdb/src/main/scala/com/twitter/zipkin/anormdb/StorageBuilder.scala
@@ -1,32 +1,0 @@
-/*
- * Copyright 2013 Twitter Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.twitter.zipkin.anormdb
-
-import com.twitter.zipkin.builder.Builder
-import com.twitter.zipkin.storage.Storage
-import com.twitter.zipkin.storage.anormdb.{AnormStorage, DB}
-
-object StorageBuilder {
-  def apply(db:DB) = {
-    new StorageBuilder(db)
-  }
-}
-class StorageBuilder(db: DB) extends Builder[Storage] {
-  def apply() = {
-    AnormStorage(db)
-  }
-}
